Convert Vue front-end to Vite
