+++ meshtastic-site-planner/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "meshtastic-site-planner",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vue-tsc -b && vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@vueuse/core": "^12.0.0",
+    "bootstrap": "^5.3.3",
+    "georaster": "^1.6.0",
+    "georaster-layer-for-leaflet": "^3.10.0",
+    "leaflet": "^1.9.4",
+    "leaflet.locatecontrol": "^0.82.0",
+    "pinia": "^2.3.0",
+    "vue": "^3.5.13",
+    "vue3-popper": "^1.5.0"
+  },
+  "devDependencies": {
+    "@types/leaflet": "^1.9.15",
+    "@types/vue": "^1.0.31",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "typescript": "~5.6.2",
+    "vite": "^6.0.1",
+    "vue-tsc": "^2.1.10"
+  }
+}
