+++ meshtastic-site-planner/src/main.ts
@@ -1,0 +1,9 @@
+import { createApp } from 'vue'
+import './style.css'
+import App from './App.vue'
+import { createPinia } from 'pinia'
+
+const app = createApp(App)
+app.use(createPinia())
+
+app.mount('#app')
