+++ meshtastic-site-planner/src/style.css
@@ -1,0 +1,20 @@
+html,
+body {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+#map {
+  height: 100vh;
+  margin: 0;
+  padding: 0;
+}
+
+.leaflet-top {
+  top: 70px; /* Adjust this value to the height of your navbar */
+}
+
+.spinner-border-sm {
+    margin-right: 0.5rem; /* Optional: Add spacing between spinner and text */
+}
